Add twig comments to link location events
These comments are added to twig location events that are reversed. Sometimes an event named "after" actually renders in a position before, because the event name refers to code location, not rendered placement. This is also inconsistent between subsilver2 and prosilver, so a comment has been added to any files where the actual event selected is not a match for the event file needed.